refactor transducers component and add home page for routing
