+++ src/App.tsx
@@ -1,14 +1,16 @@
 import { createBrowserRouter, RouterProvider } from 'react-router-dom';
 
+import MessagePage from './components/MessagePage/MessagePage';
 import Root from './pages/Root';
-import Transducers from './components/Transducers/Transducers';
+import Home from './pages/Home';
 
 const router = createBrowserRouter([
   {
     path: '/',
     element: <Root />,
+    errorElement: <MessagePage message='Page cound not be found!' />,
     children: [
-      { path: '/', element: <Transducers /> }
+      { path: '/', element: <Home /> }
     ]
   }
 ]);
